+++ composer/packages/vs-syntax-tree/src/components/diagnosticsPopup.tsx
