+++ composer/packages/vs-syntax-tree/src/components/nodeDetailsPopup.tsx
