+++ composer/packages/vs-syntax-tree/src/components/treeEdge.tsx
