+++ composer/packages/vs-syntax-tree/src/representations/dropdown-tree.tsx
@@ -1,5 +1,5 @@
 import React, {useEffect, useState} from "react";
-import { DropdownTreeProps } from "./tree-interfaces";
+import { DropdownTreeProps } from "../tree-interfaces";
 
 function DropdownTree(props: DropdownTreeProps) {
     const [isCollapsed, updateisCollapsed] = useState(false);
@@ -19,6 +19,7 @@ function DropdownTree(props: DropdownTreeProps) {
                     backgroundColor: "yellow",
                     border: "2px solid black",
                     height: 50,
+                    lineHeight: "50px",
                     width: 350
                 }}
                 onClick = {changeCollapsibleStatus}
